% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_extra_step.r
\name{Evaluate}
\alias{Evaluate}
\title{Evaluate tree}
\usage{
Evaluate(tree, dataset)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{data}{dataset of class \code{phyDat}}
}
\description{
Evaluate tree
}
