% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_extra_step.r
\name{ICPerStep}
\alias{ICPerStep}
\alias{ICS}
\title{Information content per step}
\usage{
ICPerStep(splits, maxIter, warn = TRUE)

ICS(..., envir = parent.frame())
}
\arguments{
\item{splits}{TBC}

\item{maxIter}{number of iterations to use when estimating concavity constant}

\item{warn}{Boolean (\code{TRUE}/\code{FALSE}): display warnings when concavity 
functions are generated by approximation.}

\item{a}{\code{min(splits)}.}

\item{b}{\code{max(splits)}.}

\item{m}{\code{maxIter}.}

\item{warn}{Boolean (\code{TRUE}/\code{FALSE}): display warnings when concavity 
functions are generated by approximation.}
}
\description{
Information content per step
}
\section{Functions}{
\itemize{
\item \code{ICS}: Memoized calculating function
}}

\keyword{internal}
