% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_extra_step.r
\name{InfoAmounts}
\alias{InfoAmounts}
\title{Amount of information in each character}
\usage{
InfoAmounts(dataset, precision = 1e+06, warn = TRUE)
}
\arguments{
\item{precision}{number of random trees to generate when calculating Profile curves}

\item{warn}{Boolean (\code{TRUE}/\code{FALSE}): display warnings when concavity 
functions are generated by approximation.}

\item{data}{dataset of class \code{phyDat}}
}
\value{
information content of each extra step, in bits
}
\description{
Amount of information in each character
}
\author{
Martin R. Smith
}
