% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{PhyDat}
\alias{PhyDat}
\title{Load phyDat object}
\usage{
PhyDat(dataset, levels = NULL, compress = TRUE)
}
\arguments{
\item{dataset}{data table, perhaps from read.nexus.data}

\item{levels}{tokens - values that all characters migkt take}

\item{compress}{Compress identical transformation series into a single row of the phyDat object
For simplicity I have not retained support for contrast matrices or ambiguity.}
}
\value{
a \code{phyDat} object
}
\description{
A convenient wrapper for \pkg{phangorn}'s \code{phyDat}
}
