% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{PrepareDataProfile}
\alias{PrepareDataProfile}
\title{Prepare data for Profile Parsimony}
\usage{
PrepareDataProfile(data, precision = 4e+05)
}
\arguments{
\item{data}{dataset of class \code{phyDat}}

\item{precision}{number of random trees to generate when calculating Profile curves}
}
\value{
a dataset of class 'profileDat'
}
\description{
Prepare data for Profile Parsimony
}
\author{
Martin R. Smith; written with reference to phangorn:::prepareDataFitch
}
