% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileScore.R
\name{ProfileScore}
\alias{ProfileScore}
\title{Profile Parsimony Score}
\usage{
ProfileScore(tree, dataset)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{dataset}{Dataset of class \code{profileDat} (see \code{\link{PrepareDataProfile}})
Alternatively, a dataset of class \code{phyDat} can be provided, and will 
be (time-consumingly) converted within the function.}
}
\value{
Zero minus the profile score (because the optimization algorithm assumes that
        smaller numbers are better)
}
\description{
Calculate a tree's Profile Parsimony score with a given dataset, after Faith and Trueman (2001)
}
\examples{
  data(referenceTree)
  data(congreveLamsdellMatrices)
  dataset <- PrepareDataFitch(congreveLamsdellMatrices[[42]])
  ProfileScore(referenceTree, dataset)

}
\references{
Faith, D. P. & Trueman, J. W. H. (2001). \cite{Towards an inclusive philosophy for 
   phylogenetic inference.} Systematic Biology 50:3, 331-350, doi: 
   \href{http://dx.doi.org/10.1080/10635150118627}{10.1080/10635150118627}
}
\author{
Martin R. Smith
}
\keyword{tree}
