\name{treeSearchResults}
\alias{treeSearchResults}
\docType{data}
\title{Optimal and suboptimal trees from Congreve & Lamsdell (2016) matrices}
\description{
  Optimal and suboptimal trees generated by running a parsimony ratchet (\code{\link{Pratchet}})
  on the simluated matrices from Congreve & Lamsdell (2016) (\link{congreveLamsdellMatrices}),
  using a random starting tree
}
\usage{
  data(treeSearchResults)
}
\format{
  A list of 100 \code{multiPhylo} objects, each containing a number of optimal and suboptimal 
  trees.  Each individual tree has an attribute 'score' that provides the profile parsimony 
  score.
}
\details{
  These trees were the basis of the data analysis in Smith (201X).
}
\source{
data(congreveLamsdellMatrices)
lapply(congreveLamsdellMatrices, function (data) {
  Pratchet(RandomTree(data), PrepareDataFitch(data), all=TRUE, outgroup='1', suboptimal=24,
           k=100, searchhits=50, searchiter=5000, pratchiter=5000, rearrangements='TBR')
})
\references{
\itemize{
\item Congreve, C. R., & Lamsdell, J. C. (2016). \cite{Implied weighting and its utility in 
  palaeontological datasets: a study using modelled phylogenetic matrices.} Palaeontology, 
  online ahead of print. doi:10.1111/pala.12236

\item Congreve, C. R., & Lamsdell, J. C. (2016). Data from: Implied weighting and its utility in 
palaeontological datasets: a study using modelled phylogenetic matrices. Dryad Digital
Repository, \href{http://dx.doi.org/10.5061/dryad.7dq0j}{doi:10.5061/dryad.7dq0j}

\item Smith, M. R. (201X). \cite{Trade-Offs between Resolution and Accuracy under Different Methods of
Phylogenetic Inference.} Syst. Biol. 
}
\examples{
data('treeSearchResults')
data('congreveLamsdellMatrices')
dat <- lapply(congreveLamsdellMatrices, PrepareDataFitch)
# List trees within 0.1 bits of the best score
optimal.trees <- lapply(1:100, function (i) {
  scores <- vapply(treeSearchResults[[i]], attr, double(1), 'score')
  treeSearchResults[[i]][scores <= min(scores) + 0.1]
})
optimal.consensus <- lapply(optimal.trees, consensus)  
}
}
\keyword{datasets}