% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_extra_step.r
\name{WithOneExtraStep}
\alias{WithOneExtraStep}
\title{Number of trees with one extra step}
\usage{
WithOneExtraStep(splits)
}
\arguments{
\item{splits}{TBC}
}
\description{
Number of trees with one extra step
}
