% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congreveLamsdellMatrices}
\alias{congreveLamsdellMatrices}
\title{100 simulated data matrices}
\format{A list with 100 entries, each comprising a phyDat object of 55 characters for 22 taxa}
\source{
\url{http://datadryad.org/resource/doi:10.5061/dryad.7dq0j}
}
\usage{
congreveLamsdellMatrices
}
\description{
Contains the 100 simulated matrices generated by Congreve & Lamsdell (2016) using a 
 heterogeneous Markov-k model, generated from the \link{referenceTree} topology, with 
 all branches sharing an equal length.
}
\examples{
  data('referenceTree')
  data('congreveLamsdellMatrices')
  ProfileScore(referenceTree, PrepareDataProfile(congreveLamsdellMatrices[[17]]))

}
\references{
Congreve, C. R., & Lamsdell, J. C. (2016). \cite{Implied weighting and its utility in 
  palaeontological datasets: a study using modelled phylogenetic matrices.} Palaeontology, 
  online ahead of print. doi:10.1111/pala.12236

Congreve, C. R., & Lamsdell, J. C. (2016). Data from: Implied weighting and its utility in 
palaeontological datasets: a study using modelled phylogenetic matrices. Dryad Digital
Repository, \href{http://dx.doi.org/10.5061/dryad.7dq0j}{doi:10.5061/dryad.7dq0j}
}
\keyword{datasets}
