% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_extra_step.r
\name{Evaluate}
\alias{Evaluate}
\title{Evaluate tree}
\usage{
Evaluate(tree, dataset, warn = TRUE)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{dataset}{dataset of class \code{phyDat}}

\item{warn}{Boolean (\code{TRUE}/\code{FALSE}): display warnings when concavity 
functions are generated by approximation.}
}
\description{
Evaluate tree
}
