% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{PrepareDataProfile}
\alias{PrepareDataProfile}
\title{Prepare data for Profile Parsimony}
\usage{
PrepareDataProfile(dataset, precision = 4e+05, warn = TRUE)
}
\arguments{
\item{dataset}{dataset of class \code{phyDat}}

\item{precision}{number of random trees to generate when calculating Profile curves. 
With 22 tokens (taxa), a precision increase of 4e+05 to 8e+05 sees
little difference in most steps; 80% of the differences are within 
0.05 bits, and c. 95% within 0.20 bits
values > 1e+06 consume enough memory to cause my 2012 desktop to struggle}

\item{warn}{Boolean (\code{TRUE}/\code{FALSE}): display warnings when concavity 
functions are generated by approximation.}
}
\value{
a dataset of class 'profileDat'
}
\description{
Prepare data for Profile Parsimony
}
\author{
Martin R. Smith; written with reference to phangorn:::prepareDataFitch
}
