% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{referenceTree}
\alias{referenceTree}
\title{Tree topology for matrix simulation}
\format{A single phylogenetic tree saved as an object of class \code{phylo}}
\source{
\url{http://dx.doi.org/10.1111/pala.12236}
}
\usage{
referenceTree
}
\description{
The tree topology used to generate the matrices in \code{\link{congreveLamsdellMatrices}} 
Congreve & Lamsdell (2016)
}
\examples{
  data(referenceTree)
  plot(referenceTree)

}
\references{
Congreve, C. R., & Lamsdell, J. C. (2016). \cite{Implied weighting and its utility in 
  palaeontological datasets: a study using modelled phylogenetic matrices.} Palaeontology, 
  online ahead of print. doi:10.1111/pala.12236

Congreve, C. R., & Lamsdell, J. C. (2016). Data from: Implied weighting and its utility in 
palaeontological datasets: a study using modelled phylogenetic matrices. Dryad Digital
Repository, \href{http://dx.doi.org/10.5061/dryad.7dq0j}{doi:10.5061/dryad.7dq0j}
}
\keyword{datasets}
